feat(turtle): Logo-faithful coordinates — centered origin, +Y up, WRAP
The TURTLE module ($07) reported raw screen pixels (home = (160,80)); every Logo
dialect (UCBLogo/FMSLogo/Terrapin/Apple/MIT/Python-logo-mode) puts the origin
(0,0) at screen center with +Y up. Recenter to the full-plane center (160,100)
and translate at the op boundary so the internal render/move-math core stays
screen-pixel and untouched:

- Centered frame: sx = lx+160, sy = 100-ly (and inverse) at SETXY/SETX/SETY and
  XCOR/YCOR; TOWARDS converts its target the same way. HOME now reports (0,0),
  FD north reports +YCOR, RT 90 + FD reports +XCOR. CENTER_Y 80 -> 100.
- WRAP (default edge behavior): the stored position is reduced mod 320x200, and a
  pen line that crosses an edge is drawn wrapped via tile-shift (draw the line
  shifted by each plane-tile it spans; the VGC clips each copy).
- TOWARDS now returns a Logo heading (north=0, east=90, clockwise) via
  heading = (90 - mathdeg) mod 360, so SETHEADING TOWARDS [x y] points at the
  target instead of ~90 degrees off.
- VGC line endpoints are signed-16 + Cohen-Sutherland clipped to the plane
  (the correct layer for the WRAP tile-shift) — also fixes a latent bug where a
  negative endpoint read as ~65000 and produced a garbage, slow line for every
  caller (GRAPHICS, BASIC).

The Logo adapter passes numbers through unchanged, so the whole convention lives
in the module. Full test suite green (1558 pass): 3 new VGC tests, 5 coordinate
contract tests, 3 WRAP tests, TOWARDS rewritten to assert all 4 cardinals, and
the existing turtle/Logo tests retargeted to the new convention.

Design: docs/plans/2026-06-06-turtle-logo-coordinates-design.md

+++ e6502.Avalonia/Hardware/VirtualGraphicsController.cs
@@ -987,7 +987,11 @@ public class VirtualGraphicsController : IBusDevice
                 BlockGraphics.Plot(_gfxBitmap, p0, p1, 0);
                 break;
             case VgcConstants.CmdLine:
-                BlockGraphics.Line(_gfxBitmap, p0, p1, p2, p3, color);
+                // Line endpoints are signed 16-bit: the turtle WRAP path draws a
+                // line shifted by whole plane-tiles and relies on the off-plane
+                // copies clipping cleanly (a negative endpoint must mean "left/above
+                // the plane", not ~65000). BlockGraphics.Line clips to the plane.
+                BlockGraphics.Line(_gfxBitmap, (short)p0, (short)p1, (short)p2, (short)p3, color);
                 break;
             case VgcConstants.CmdCircle:
                 int ry = _cmdRegs[7] | (_cmdRegs[8] << 8);
